Convert float to int before unsigned
setreg takes an unsigned value argument since some uses such as
durec use 16-bit values with the high bit set. Any calls with an
int argument work as expected since the int to unsigned conversion
uses two's complement. However, a negative float argument can't be
represented by unsigned, so we must first convert to int.

This happened to work on x86_64, but broke for the wasm build.

+++ oscmix.c
@@ -179,7 +179,7 @@ setfixed(const struct oscnode *path[], int reg, struct oscmsg *msg)
 	val = oscgetfloat(msg);
 	if (oscend(msg) != 0)
 		return -1;
-	setreg(reg, val / node->scale);
+	setreg(reg, (int)(val / node->scale));
 	return 0;
 }
